refactor: rename saloon to salon, use uuids, and rename entity files to schema
