+++ src/app/user/infrastructure/schemas/user.schema.ts
