+++ src/config/typeorm.config.ts
@@ -1,6 +1,6 @@
 import { DataSource } from 'typeorm';
 import * as dotenv from 'dotenv';
-import { User } from '@app/user/infrastructure/schemas/user.entity';
+import { User } from '@app/user/infrastructure/schemas/user.schema';
 import { Salon } from '@app/salon/infrastructure/schemas/salon.entity';
 import { SalonMembershipEntity } from '@app/user/infrastructure/schemas/salon-membership.schema';
